Use uboot.env.
uEnv.txt is apparently not supported by the Rpi part of u-boot. In addition
the more common use of boot.scr also fails, because for some reason loading it
at address 0 modifies the first bit, causing it not to recognise the image
format. If you manually modify the first bit to the right type, it fails to
boot. So boot.scr is out of the question. But this works.